requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
protobuf==4.23.2
psycopg2-binary==2.9.6
